Added database reference file.
Added file sql-table-structure-reference.txt, which is a reference file on how
the database will probably look.  As the comments explain in the beginning of
the file, the current structure of the database is not reflected by this file,
some tables still need to be fleshed out, and some of the tables need to be
revised further.  Additionally, this file contains some views and custom
functions that have been useful working with the data so far.