Added initial vesion of a schema file.
Added a paps.sql schema file.  This defines the current version of the
database, which right now is comprised of four tables with some foreign keys
relating them.  When this script is run it would first drop any of these
tables, if they exist, to make sure we do not try to recreate something that
already exists.