Added wrapper TT file to provide general site content.
Created a TT file wrapper.tt2 that acts as a wrapper around the content on
each page.  Also added this file as the wrapper file in the TT.pm view module.

The wrapper file right now has only very basic content taken from the tutorial;
it will need to be updated to better suit the PAPS site.