Added many-to-many relationship between works and authors.
Added the two ends of the many-to-many relationship between works and authors
in their respective schema classes.  This many-to-many relationship is an
extension of the has_many and belongs_to relationships that were automatically
created from the database structure.  The many-to-many relationship is not
required, but it makes accessing, for example, a work's authors, easier.