Added a list action and view for People controller.
Added a local action "list" to the People controller.  This is very similar to
the list action for the Works controller.  Also added a people/list.tt2 view,
which lists the people in the db.  It gets any works a person is associated
with and lists them.  The algorithm to get and list the works is inefficient
and misplaced, but is here for testing and demonstration purposes.