Renaming application 1: Moving files.
First part in an effort to rename the application.  This is not done using any
fancy script or anything; I am simply recreating the application with a new
name and reapplying changes.  Name change is small and almost to the point of
being silly: changing from 'paps' to 'PAPS', primarily to align with Perl
module naming standards.

This step includes file renames due to having a new module name.