Changed DOI column to be a hyperlink.
Added a proper DOI link prefix before the DOI value, and turned the entire
string into a hyperlink.  There are probably a bunch of different "base" URIs
to use for DOI values, but this will get us started.  Also, there are probably
much better ways to turn this into a link, but, again, this works for now.