Added functionality to create a new work.
Added a new controller action create_form ("create" a new work using a "form")
that maps to the path "/works/new".  This action loads a new template
"new.tt2" that presents a form the user can fill out and submit.  After
submitting, the user is presented with the "create_done.tt2" template, which
gives a little information from the work they just created.  The created work
shows up in the list of works when the user navigates there.