Replaced logic to get author list with call to author_list.
The previous logic to get the list of authors was inefficient and had all of
the code inside the TT file, which is the wrong location for it.  Replaced this
logic with a call to the Works schema helper method author_list, which does all
the necessary lookup and formatting.