Added functionality to create new People.
Added new Controller methods create_form and do_create that handle showing a
form to create new People and add the new Person to the database, respectively.
Also replaced the index method with a base method that can be chained off of.
Added new TT pages to provide a form for the new Person's information and
show the new Person was created.  These additions are based off of what has
been done for the Works controller/model.