Reference: Redefining a has_many relationship.
The DBIx::Class::Schema::Loader class automatically creates a set of Schema
classes based on a database.  These, however, are very basic and most likely
need to be modified later on.  The commit here shows we can "redefine" a
relationship (here a has_many relationship) with different properties.  In the
redefinition here, there are multiple examples about how to set attributes,
such as proxies or additional fields to include in the result set.  These
attributes are included in a hash ref in the fourth parameter to the has_many
method.  In this specific case, these examples do not work as intended (add the
author_position field to the authors result set), but they can be used as
reference.