Added author_position to many_to_many relationship.
Modified the definition for the many_to_many relationship between the works and
authors tables to include the work_authors.author_position field in the select
list.  This makes it much easier (and more efficient) to get the position for
each author of a particular work, starting from the work row/object.