Added reference_types and work_references tables.
Added schema information for the reference_types and work_references tables.
Also added sample records to be inserted into both of these tables.  The
work_references table was originally going to be named references, but that
word is a keyword in SQL, so it was causing syntax errors when trying to
create the tables and relationships.