Added lists for referenced and referencing works.
Added additional tables to the work details page to list the works this work
references and the works that reference this work.  When the other work is
known (which is always the case for referencing works, but not always for
referenced works), the work's name is a link to that work's details page.