Added reference id as an argument for do_edit_reference.
Included an additional parameter to supply arguments for the do_edit_reference
action in the Works controller.  This action is chained off of other actions,
so it also needs an array ref containing the id of the Work in question.  The
id of the reference must be in a different argument, as this is an argument to
that action, not one of the captures (arguments used in parts of the chain
that are not endpoints).