Changed referencing_work list to use has-many relationship.
The list of Works referencing the given Work was using the referencing_works
many-to-many relationship, which provided the actual Work objects.  This made
getting all of the columns from the linking table (WorkReferences) more
difficult, particularly getting the name of the type of the reference.  This
list now uses the work_references_referencing_works has-many relationship.