Added reference type to details and edit views.
Added a column on the Work details page with the type of the Reference in the
References and Cited By lists.  In the edit page, added a column with a
dropdown box with the list of Reference Types.  For existing References, the
current Reference Type is selected.  For new References, a blank value is
selected, but this must be changed, as the reference_type_id column is not
nullable.