Added graph action to create a graph of Works.
Added a simple acton 'graph' that uses GraphViz to create a simple graph of
the Works in the database and the References between them.  This action
creates the graph each time the action is executed, so it is not too efficient
right now.  The graph is saved to the root/static/images directory.