Added actions and views to create new sources.
Added create_form ('new' path) and do_create actions to present the user with
a form to create a new Source and to inform him the Source was created,
respectively.  The new.tt2 view provides a table of input fields for the new
Source.  The create_done.tt2 view simply provides some information about the
Source that was just created.