Changed overlap setting for GraphViz Works graph.
Changed the overlap setting for the GraphViz graph to 'scalexy', from 'false'.
The 'false' setting had the nodes not overlap, but the resulting graph was
fairly large.  The 'scalexy' setting, instead, scales the nodes so they do not
overlap, and is more conservative in how large it makes the graph.