Added action to list all References.
The list action in the References controller stores all of the WorkReference
objects in the stash.  These objects are sorted by the referencing_work_id,
chapter, and rank.  Currently, we can not sort by the title or display name of
the referencing works, as that would require joining on the works table for
each Reference.  Will need to optimize these queries to be able to do that.