Added create, list, and edit actions and views for SourceCategoryTypes.
Added controller actions and views for the SourceCategoryTypes controller to
create new source category types, list all source category types, and edit the
properties for source category types (this does not support editing the source
categories in a source category type).  Also changed the base url for the
controller to "sourcecategorytypes".