Added tag type to works' list of source tags.
Added a "Tag Type" column to the list of source tags for a work.  Also rewrote
the query to fetch the work's source tags and related metadata.  Instead of
just using the many-to-many relationship, perform a serch on the has-many
relationship and join on the relationships to the source_tags, sources, and
source_tag_types tables.  This allows us to include arbitrary fields from
these tables and order by these fields.