Added graph2 action, which uses GraphViz2 package to graph works.
The GraphViz2 package is a newer package that allows drawing graphs with
GraphViz.  Added a new action, graph2, that uses this package to create a
graph of all of the works and their reference relationships.  Each node
includes additional information, including the works' authors, categories, and
tags.  Since this requires accessing additional tables, this action takes
substantially more time and resources to complete.