Added controller action and views to edit a User.
Since passwords are handled uniquely, it is not easy to edit a user's password
from the same edit form as all his other properties (name, email, etc.).  As
a result, it is not currently possible to change a user's password.  This can
be added later in a special "change password" action.  Also, enabling and
disabling a user is not supported from this edit page.  This could be added
later or could be handled from another kind of page.