Added Users controller actions for users to log in.
Added a 'login' action to the Users controller that loads the login view, to
allow the user to enter user information.  Also added a 'do_login' action that
checks if the provided information matches a vaild user and has a valid,
matching password.  Debug information relating to how the login attempt has
resolved is printed to the console.