Added module dependences for session support.
Session support requires two parts, a way to save the session state and a way
to store the session information.  The former is typically done using cookies,
which is how it is stored here.  The session store can be much more variable.
For now, we will be going with a simple FastMmap implementation, which stores
session information in a file on disk.  This does not support serving the web
application across multiple serves, but we will cross that bridge when we come
to it.