Added action to create or edit the user's work data.
Added the do_edit_user_data action to the Works controller.  This action
creates or updates the currently-logged-in user's data for the work being
edited.  If the user does not have any data for the work (specifically, if no
such data was stored in the stash), a record is created with the data that was
entered by creating the appropriate related object.  If the user does have
data for the work, it is updated.