Before fetching work info to graph, add logged in user's data.
The graph2 action graphs all of the works and their relationships.  If a user
is logged in, we want to add his user information to the nodes.  This data is
fetched by adding a join on the user_work_datas relationship and a where
clause to limit the data to the current user's information.