Added explicit dependency on the papsdb Schema submodule.
While it is likely not possible to successfully install either this module or
the PAPS::Database::papsdb::Schema module, I am including a reference to the
Schema module as an explicit dependency here.  This is mostly to note the
minimum required version; the current version of the PAPS module relies on
features that are only in version 0.005 of the papsdb schema module.