Changed "disabled" attributes to "readonly."
The "disabled" attribute prevents the user from editing an input element, but
it also prevents that value from being submitted in the form.  This is not an
acceptable situation, so the input elements in the user's edit collection
table have been changed to use "readonly" attributes, instead.  This also has
the side-benefit of allowing the user to select the text in those fields, for
example so he can copy them.