Added new method create_graph to create a graph.
This is an attempt to generalize how graphs are created.  The new create_graph
method takes several parameters that specify what is being graphed, and how it
is being graphed.  The $works_rs object is the result set containing the works
that should be graphed, and the $ref_rs object is the result set containing
the references that should be graphed.  Each of these result sets should have
the necessary columns to perform this graphing (this is not error-proof).  For
example, if a user is logged in, the works_rs object should have columns for
the current user's metadata about the works.  This method returns the file
name (not the path) to the file of the graph that was created.