Updated create_work_graph method to use create_graph method.
The create_work_graph method previously created a graph for a work and its
references (both what it reference and what referenced it).  This required two
searches to get all of the references, as well as additional queries to get
work information for works other than the initial one.  This new version of
the method uses the generic create_graph method to do the heavy lifting for
actually creating the GraphViz graph.  Instead of the aforementioned searches,
it uses the id of the work passed in to find all works referencing it or
referenced by it, and then all related references, and passes these result
sets to the create_graph method.