Added method create_work_reading_graph.
Here, a "reading graph" is being defined as the graph of works and references
between works for a work that indicate what works need to be read before one
can expect to be able to understand that work.  If a person wants to be
"safe," he can read all of thw works first.  If not, he can skip some of the
works lower in the tree.  Later, additional colors wiil be added to help the
person easily see the status or state of each work in the graph.