Collect information on each node for the reading graph.
After creating the search result set for works, iterate through it and get the
number of times each work references another work and how many times a work is
referenced.  Store these in hashes, and use the set of keys for the latter to
get the list of works to search for references in.