Moved checking for not understood earlier and added checks for no ratings.
In the reading graph coloring scheme, moved checking if a work is not
understood to before checking if it has no references.  That is, treat a work
is a bad choice if any of its references could not be understood.  Also, some
of the checks were not behaving properly if none of the references had been
rated, so I added a check to see if the minimum understanding rating existed
(as a null/default value is indeed less than the understanding threshold).