Only update a work's min understanding if, additionally, the reference's understanding is defined.
When calculating a work's minimum undersood rating for the reading graph, the
test to update the minimum understood rating did not check if the reference in
question had a defined understood rating.  Thus, since the null/default value
was less than the current minimum, the minimum value was set to the
null/default value, which caused nodes to be colored incorrectly.