Updated graph2 action to use create_graph method.
The graph2 action previously handled all aspects of creating the works graph.
To improve code reuse, this now uses the create_graph method.  The only
outwardly-visible difference is that the resulting graph uses the default
layout format.  This can be fixed down the road by specifying the layout
format in the settings hash.