Use get_column method to get values from joins.
Columns that are part of a result set due to a join can not be accessed using
the simple arrow operator (accessing a hash key), as far as I am aware.  For
the current consumers of the determine_node_counts_and_reference_counts method,
this is fine, as the columns such as understood_rating and referencing_work_id
were part of the virtual view.  The graph2 action, however, would encounter
problems, as it gets these columns using joins.

To fix this, each of the columns not part of the Work result set are now
accessed using the get_column method.