Calculate reading data in graph2 action to colorize the graph.
In the graph2 action, call the determine_node_counts_and_reference_counts and
determine_work_custom_fill_colors methods to determine, based on the current
user's work data and the work references, how to color each node.  These
colors are the same as the reading graphs and the connected component graphs.