Do not fetch user work data in graph2 action if there is no user.
The search on the Work result set, when no user was logged in, was only
returning records for works that had no user work data.  Thus, works that had
user work data were not getting proper nodes in the resulting graph.  This
would be very difficult to fix in one search, so different searches are not
set up for when a user is logged in and for when no user is logged in.  When
no user is logged in, there is no where clause for the user work data, and
instead of pulling the columns from the user_work_data table, simply null
values are used (literal "null"s are injected into the query using string
references).  These null values are then handled in the following methods, as
they are undefined values.