Use ViewWorksWithUserData view in graph2 action.
The search used in the graph2 action to get all works with the current user's
user work data did not return works that the current user did not have work
data for but other users did, due to how the query and tables were
constructed.  The ViewWorksWithUserData view was created to support getting
all works with the current user's (if any) user work data.